Changes in batch band example
