+++ examples/batched-band-solver/batched-band-solver.cpp
@@ -58,6 +58,48 @@ using batch_csr = gko::matrix::BatchCsr<value_type>;
 using batch_band_solver = gko::solver::BatchBandSolver<value_type>;
 
 
+// @sect3{'Application' structures and functions}
+// Structure to simulate application data related to the linear systems
+// to be solved.
+//
+// We use raw pointers below to demonstrate how to handle the situation when
+// the application only gives us raw pointers. Ideally, one should use
+// Ginkgo's gko::Array class here.
+struct ApplSysData {
+    // Number of small systems in the batch.
+    size_type nsystems;
+    // Number of rows in each system.
+    int nrows;
+    // Number of sub-diagonals in each system matrix.
+    int num_sub_diags;
+    // Number of  super-diagonals in each system matrix.
+    int num_super_diags;
+    // Column major batch band array (in the prescribed fashion)
+    const value_type* batch_band_array;
+    // RHS vectors for all systems in the batch, concatenated
+    const value_type* all_rhs;
+};
+
+// Generates a batch of tridiagonal systems.
+//
+// @param nrows  Number of rows in each system.
+// @param num_sub_diags Number of sub-diagonals in each system matrix
+// @param num_super_diags  Number of  super-diagonals in each system matrix
+// @param nsystems  Number of systems in the batch.
+// @param exec  The device executor to use for the solver.
+//   Normally, the application may not deal with Ginkgo executors, nor do we
+//   need it to. Here, we use the executor for backend-independent device
+//   memory allocation. The application, for example, might assume Hip
+//   (for AMD GPUs) and use `hipMalloc` directly.
+ApplSysData appl_generate_system(const int nrows, const int num_sub_diags,
+                                 const int num_super_diags,
+                                 const size_type nsystems,
+                                 std::shared_ptr<gko::Executor> exec);
+
+// Deallocate application data.
+void appl_clean_up(ApplSysData& appl_data, std::shared_ptr<gko::Executor> exec);
+
+
 int main(int argc, char* argv[])
 {
     // Print the ginkgo version information.
@@ -65,8 +107,7 @@ int main(int argc, char* argv[])
 
     if (argc == 2 && (std::string(argv[1]) == "--help")) {
         std::cerr << "Usage: " << argv[0]
-                  << " [executor] [dir_path] [problem_name] [num_systems] "
-                     "[num_duplications] "
+                  << " [executor] [N] [KL] [KU] [num_systems] "
                      "[band solver approach] [blocked solve panel size] "
                      "[file_timings]"
                   << std::endl;
@@ -108,17 +149,14 @@ int main(int argc, char* argv[])
     const auto exec = exec_map.at(executor_string)();  // throws if not valid
 
     // @sect3{Other optional command-line arguments}
-    const std::string dir_name = argc >= 3 ? argv[2] : "data/";
+    const int nrows = argc >= 3 ? std::atoi(argv[2]) : 15;
 
-    // Name of the problem
-    const std::string problem_name =
-        argc >= 4 ? argv[3] : "banded_N_100_KL_7_KU_8";
+    const int KL = argc >= 4 ? std::atoi(argv[3]) : 3;
 
-    // Number of linear systems to read from files.
-    const size_type num_systems = argc >= 5 ? std::atoi(argv[4]) : 2;
+    const int KU = argc >= 5 ? std::atoi(argv[4]) : 4;
 
-    // Number of times to duplicate whatever systems are read from files.
-    const size_type num_duplications = argc >= 6 ? std::atoi(argv[5]) : 3;
+    // Number of linear systems
+    const size_type num_systems = argc >= 6 ? std::atoi(argv[5]) : 2;
 
     // Approach
     enum gko::solver::batch_band_solve_approach approach;
@@ -135,47 +173,48 @@ int main(int argc, char* argv[])
 
     const char* log_file = argc >= 9 ? argv[8] : "timings_file.txt";
 
-    // @sect3{Read batch from files}
-    auto data = std::vector<gko::matrix_data<value_type>>(num_systems);
+    // @sect3{Generate data}
+    // The "application" generates the batch of linear systems on the device
+    auto appl_sys = appl_generate_system(nrows, KL, KU, num_systems, exec);
+    // Create batch_dim object to describe the dimensions of the batch matrix.
+    auto batch_mat_size =
+        gko::batch_dim<>(num_systems, gko::dim<2>(nrows, nrows));
+    auto batch_num_subdiags = gko::batch_stride(num_systems, KL);
+    auto batch_num_superdiags = gko::batch_stride(num_systems, KU);
+    auto batch_vec_size = gko::batch_dim<>(num_systems, gko::dim<2>(nrows, 1));
+    // @sect3{Use of application-allocated memory}
+    // We can either work on the existing memory allocated in the application,
+    //  or we can copy it for the linear solve.
+    //  Note: it is not possible to use data through a const pointer directly.
+    //  Because our pointers are not const, we can just 'wrap' the given
+    //  pointers into Ginkgo Array views so that we can create a Ginkgo matrix
+    //  out of them.
+    // Ginkgo expects the nonzero values for all the small matrices to be
+    //  allocated contiguously, one matrix after the other.
 
-    for (size_type i = 0; i < data.size(); ++i) {
-        const std::string mat_str = "A.mtx";
-        const std::string fbase =
-            dir_name + problem_name + "/" + std::to_string(i) + "/";
-        std::string fname = fbase + mat_str;
-        std::cout << "\nReading file: " << fname << std::endl;
-        std::ifstream mtx_fd(fname);
-        data[i] = gko::read_raw<value_type>(mtx_fd);
-    }
+    auto band_arr_vals_view = gko::array<value_type>::const_view(
+        exec, num_systems * nrows * (2 * KL + KU + 1),
+        appl_sys.batch_band_array);
 
-    auto single_batch = mtx_type::create(exec);
-    single_batch->read_band_matrix(data);
-    auto single_batch_csr = batch_csr::create(exec);
-    single_batch_csr->read(data);
+    auto A = gko::share(mtx_type::create_const(
+        exec, batch_mat_size, batch_num_subdiags, batch_num_superdiags,
+        std::move(band_arr_vals_view)));
 
-    // We can duplicate the batch a few times if we wish.
-    std::shared_ptr<mtx_type> A =
-        mtx_type::create(exec, num_duplications, single_batch.get());
-    std::shared_ptr<batch_csr> A_csr =
-        batch_csr::create(exec, num_duplications, single_batch_csr.get());
+    std::vector<gko::matrix_data<value_type>> data;
+    A->write(data);
+    auto A_csr = gko::share(batch_csr::create(exec));
+    A_csr->read(data);
 
-    // Create RHS
-    const auto nrows = A->get_size().at(0)[0];
-
-    const size_type num_total_systems = num_systems * num_duplications;
-
-    auto host_b = vec_type::create(
-        exec->get_master(),
-        gko::batch_dim<2>(num_total_systems, gko::dim<2>(nrows, 1)));
-
-    for (size_type isys = 0; isys < num_total_systems; isys++) {
-        for (int irow = 0; irow < nrows; irow++) {
-            host_b->at(isys, irow, 0) = static_cast<value_type>(rand());
-        }
-    }
-
-    auto b = vec_type::create(exec);
-    b->copy_from(host_b.get());
+    // @sect3{RHS and solution vectors}
+    // batch_stride object specifies the access stride within the individual
+    //  matrices (vectors) in the batch. In this case, we specify a stride of 1
+    //  as the common value for all the matrices.
+    auto batch_vec_stride = gko::batch_stride(num_systems, 1);
+    // Create RHS, again reusing application allocation
+    auto b_view = gko::array<value_type>::const_view(exec, num_systems * nrows,
+                                                     appl_sys.all_rhs);
+    auto b = vec_type::create_const(exec, batch_vec_size, std::move(b_view),
+                                    batch_vec_stride);
 
     auto x = vec_type::create(exec, b->get_size());
     auto x_dense_direct_solver = vec_type::create(exec, b->get_size());
@@ -264,21 +303,83 @@ int main(int argc, char* argv[])
 
     auto x_host = gko::share(gko::clone(exec->get_master(), x.get()));
     auto x_dd_host = gko::share(gko::clone(exec->get_master(), x.get()));
-    // for (int i = 0; i < num_total_systems; i++) {
-    //     std::cout << "Batch idx: " << i << std::endl;
-    //     for (int k = 0; k < nrows; k++) {
-    //         std::cout << "x[" << k
-    //                   << "]: " << x_host->get_const_values()[i * nrows + k]
-    //                   << "   " << x_dd_host->get_const_values()[i * nrows +
-    //                   k]
-    //                   << std::endl;
-    //     }
-    // }
+    for (int i = 0; i < num_systems; i++) {
+        std::cout << "Batch idx: " << i << std::endl;
+        for (int k = 0; k < nrows; k++) {
+            std::cout << "x[" << k
+                      << "]: " << x_host->get_const_values()[i * nrows + k]
+                      << "   " << x_dd_host->get_const_values()[i * nrows + k]
+                      << std::endl;
+        }
+    }
 
     std::ofstream timings_file;
     timings_file.open(log_file, std::ofstream::app);
-    timings_file << num_total_systems << "  " << av_time_millisec << " \n";
+    timings_file << num_systems << "  " << av_time_millisec << " \n";
     timings_file.close();
 
     return 0;
+}
+
+ApplSysData appl_generate_system(const int N, const int KL, const int KU,
+                                 const size_type nsystems,
+                                 std::shared_ptr<gko::Executor> exec)
+{
+    const int nrows_band_format = 2 * KL + KU + 1;
+    const int ncols_band_format = N;
+    std::ranlux48 rgen(15);
+    std::normal_distribution<real_type> distb(0.5, 0.1);
+
+    std::vector<value_type> batch_band_array(
+        nrows_band_format * ncols_band_format * nsystems,
+        gko::zero<value_type>());
+
+    for (size_type isys = 0; isys < nsystems; isys++) {
+        const auto batch_entry_offset =
+            isys * ncols_band_format * nrows_band_format;
+
+        for (int col_dense_layout = 0; col_dense_layout < N;
+             col_dense_layout++) {
+            const int row_dense_layout_start =
+                std::max(int{0}, col_dense_layout - KU);
+            const int row_dense_layout_end_inclusive =
+                std::min(N - 1, col_dense_layout + KL);
+
+            for (int row_dense_layout = row_dense_layout_start;
+                 row_dense_layout <= row_dense_layout_end_inclusive;
+                 row_dense_layout++) {
+                const int row_band_layout =
+                    KL + KU + row_dense_layout - col_dense_layout;
+                const int col_band_layout = col_dense_layout;
+                const value_type val = distb(rgen);
+                batch_band_array[batch_entry_offset + row_band_layout +
+                                 col_band_layout * nrows_band_format] = val;
+            }
+        }
+    }
+
+    std::vector<value_type> allb(N * nsystems);
+    for (size_type isys = 0; isys < nsystems; isys++) {
+        const value_type bval = distb(rgen);
+        std::fill(allb.begin() + isys * N, allb.begin() + (isys + 1) * N, bval);
+    }
+
+    value_type* const batch_band_arr = exec->alloc<value_type>(
+        nsystems * nrows_band_format * ncols_band_format);
+    exec->copy_from(exec->get_master().get(),
+                    nsystems * nrows_band_format * ncols_band_format,
+                    batch_band_array.data(), batch_band_arr);
+
+    value_type* const all_b = exec->alloc<value_type>(nsystems * N);
+    exec->copy_from(exec->get_master().get(), nsystems * N, allb.data(), all_b);
+
+    return {nsystems, N, KL, KU, batch_band_arr, all_b};
+}
+
+void appl_clean_up(ApplSysData& appl_data, std::shared_ptr<gko::Executor> exec)
+{
+    // In general, the application would control non-const pointers;
+    //  the const casts below would not be needed.
+    exec->free(const_cast<value_type*>(appl_data.batch_band_array));
+    exec->free(const_cast<value_type*>(appl_data.all_rhs));
 }
