+++ examples/batched-band-solver/doc/short-intro
@@ -1,1 +1,1 @@
-Using a batched band solver with matrices stored in files.
+Using a batched band solver
